requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
config==0.5.1
cryptography==37.0.4
cx-Oracle==8.3.0
et-xmlfile==1.1.0
Flask==2.2.2
h11==0.13.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
logging-config==1.0.4
MarkupSafe==2.1.1
mechanize==0.4.8
numpy==1.21.6
opencv-python==4.6.0.66
openpyxl==3.0.10
oracledb==1.0.3
outcome==1.2.0
Pillow==9.2.0
ping3==4.0.3
prettytable==3.4.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
PyPDF2==2.10.3
pyperclip==1.8.2
pyping==0.0.6
PySocks==1.7.1
python-ldap==3.4.2
python-magic==0.4.27
requests==2.28.1
rethinkdb==2.4.9
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tcping==0.1.1rc1
tcpping==0.2
termcolor==1.1.0
thinker==1.1.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.11
was==0.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wikipedia==1.4.0
wsproto==1.1.0
zipp==3.8.1
